Limit interval msg limit clear

+++ src/cogs/commands/mod/clean.py
@@ -14,12 +14,14 @@ class Clean(commands.Cog):
     @commands.command(name='clean', aliases=['clear', 'clearchat', 'cleanchat', 'cc'])
     @commands.has_permissions(manage_messages=True)
     async def clean(self, ctx, interval: int):
-        await ctx.message.delete()
-        await ctx.channel.purge(limit=interval)
-        print(f"[log] {self.bot.get_guild(ctx.message.guild.id).name} | {ctx.author.name} has cleared the chat | {date.today()}")
-        await ctx.send(embed=Embed(description=f"I have cleared {interval} messages... I can see my reflection now",
-                                   color=0x2F3136)
-                       .set_image(url=f"{await get_random_gif_by_theme('anime wash')}"), delete_after=8.0)
+        if interval <= 200:
+            await ctx.message.delete()
+            await ctx.channel.purge(limit=interval)
+            print(f"[log] {self.bot.get_guild(ctx.message.guild.id).name} | {ctx.author.name} has cleared the chat | {date.today()}")
+            await ctx.send(embed=Embed(description=f"I have cleared {interval} messages... I can see my reflection now",
+                                       color=0x2F3136)
+                           .set_image(url=f"{await get_random_gif_by_theme('anime wash')}"), delete_after=8.0)
+        else: await ctx.send("It's too big, isn't it ?", delete_after=5.0)
 
     @clean.error
     async def clean_error(self, ctx, error):
